Lower log level for message  `DUT type file doesn't exist`
<!--
Please make sure you've read and understood our contributing guidelines;
https://github.com/sonic-net/SONiC/blob/gh-pages/CONTRIBUTING.md

Please provide following information to help code review process a bit easier:
-->
### Description of PR
<!--
- Please include a summary of the change and which issue is fixed.
- Please also include relevant motivation and context. Where should reviewer start? background context?
- List any dependencies that are required for this change.
-->

Summary:
This PR is to lower the log level for the message `DUT type file doesn't exist` from **ERROR** to **INFO**.
The log message is annoying in the test result, and the log message is not monitored.
![image](https://github.com/user-attachments/assets/a5f95907-340f-4280-a400-46a1475ed932)

### Type of change

<!--
- Fill x for your type of change.
- e.g.
- [x] Bug fix
-->

- [ ] Bug fix
- [x] Testbed and Framework(new/improvement)
- [ ] Test case(new/improvement)

### Back port request
- [ ] 202012
- [ ] 202205
- [ ] 202305
- [ ] 202311
- [ ] 202405

### Approach
#### What is the motivation for this PR?
This PR is to lower the log level for the message `DUT type file doesn't exist` from **ERROR** to **INFO**.

#### How did you do it?
Change the log level.

#### How did you verify/test it?
The change is verified by running a test on a physical testbed.

#### Any platform specific information?
No.

#### Supported testbed topology if it's a new test case?
Not a new test case.

### Documentation
<!--
(If it's a new feature, new test case)
Did you update documentation/Wiki relevant to your implementation?
Link to the wiki page?
-->