[action] [PR:21072] Fix ipfwd assert assumption for TH5 (#798)
<!--
Please make sure you've read and understood our contributing guidelines;
https://github.com/sonic-net/SONiC/blob/gh-pages/CONTRIBUTING.md

Please provide following information to help code review process a bit easier:
-->
### Description of PR
<!--
- Please include a summary of the change and which issue is fixed.
- Please also include relevant motivation and context. Where should reviewer start? background context?
- List any dependencies that are required for this change.
-->

Summary: Fix ipfwd assert assumption for Broadcom TH5

The test limits the number of nhops that is used for testing. However,
the general assert assumes this limit does not exist and asserts with
CRM to make sure nhops available after the test is 0.

This causes issues on TH5 because the number of nhop groups supported
is 4095, where the limit of the test is 1024. This means that the number
of nhop groups after the test would not be 0 `(4095 - 1024 - (extra
groups) >> 0)`.

The Cisco-case of the test has the proper assert that adjusts for this
test limit. Fixing the test by including Broadcom TH5 devices to use that
proper assert.

Fixes # (issue)

### Type of change

<!--
- Fill x for your type of change.
- e.g.
- [x] Bug fix
-->

- [x] Bug fix
- [ ] Testbed and Framework(new/improvement)
- [ ] New Test case
 - [ ] Skipped for non-supported platforms
- [ ] Test case improvement

### Back port request
- [ ] 202205
- [ ] 202305
- [ ] 202311
- [ ] 202405
- [ ] 202411
- [x] msft-202412
- [x] 202505

### Approach
#### What is the motivation for this PR?
ipfwd tests failing for Arista TH5 devices.

#### How did you do it?
Adjust the assert conditions for Arista TH5 devices.

#### How did you verify/test it?
Test no longer fails on Arista TH5 devices.

#### Any platform specific information?
Broadcom TH5

#### Supported testbed topology if it's a new test case?

### Documentation
<!--
(If it's a new feature, new test case)
Did you update documentation/Wiki relevant to your implementation?
Link to the wiki page?
-->